chore: narrow analyzer excludes and split runtime complexity tests
