+++ tests/codacy-config.test.js
@@ -4,24 +4,67 @@ const fs = require('node:fs');
 const path = require('node:path');
 
 const rootDir = path.join(__dirname, '..');
+const expectedGeneratedExclusions = new Set([
+  'coverage/**',
+  'coverage-100/**',
+  'dist/**',
+  'node_modules/**',
+  '.tmp*/**',
+  'package-lock.json',
+  'npm-shrinkwrap.json',
+  'pnpm-lock.yaml',
+  'yarn.lock',
+]);
+const forbiddenScopePatterns = ['tests/**', 'docs/**', 'scripts/**', '.github/workflows/**', 'src/**'];
 
 function readRepoFile(relativePath) {
   return fs.readFileSync(path.join(rootDir, relativePath), 'utf8');
 }
 
-test('codacy repo config excludes non-product paths that vendor defaults misclassify', () => {
+function parseCodacyExcludePaths(configText) {
+  const lines = configText.split(/\r?\n/);
+  const exclusions = [];
+  let inExcludeSection = false;
+
+  for (const line of lines) {
+    if (!inExcludeSection) {
+      if (/^exclude_paths:\s*$/.test(line.trim())) {
+        inExcludeSection = true;
+      }
+      continue;
+    }
+
+    if (/^-\s+/.test(line.trim())) {
+      exclusions.push(line.trim().replace(/^-+\s+/, ''));
+      continue;
+    }
+
+    if (line.trim() !== '' && !line.startsWith(' ')) {
+      break;
+    }
+  }
+
+  return exclusions;
+}
+
+test('codacy repo config excludes only generated/build/cache artifacts', () => {
   const codacyPath = path.join(rootDir, '.codacy.yaml');
   assert.equal(fs.existsSync(codacyPath), true, '.codacy.yaml should exist');
 
   const codacyConfig = readRepoFile('.codacy.yaml');
   assert.match(codacyConfig, /^exclude_paths:/m);
-  assert.match(codacyConfig, /- coverage\/\*\*/);
-  assert.match(codacyConfig, /- coverage-100\/\*\*/);
-  assert.match(codacyConfig, /- dist\/\*\*/);
-  assert.match(codacyConfig, /- package-lock\.json/);
-  assert.doesNotMatch(codacyConfig, /- tests\/\*\*/);
-  assert.doesNotMatch(codacyConfig, /- tests\/fixtures\/\*\*/);
-  assert.doesNotMatch(codacyConfig, /- docs\/\*\*/);
-  assert.doesNotMatch(codacyConfig, /- scripts\/\*\*/);
-  assert.doesNotMatch(codacyConfig, /- \.github\/workflows\/\*\*/);
+
+  const exclusions = parseCodacyExcludePaths(codacyConfig);
+  assert.deepEqual(
+    new Set(exclusions),
+    expectedGeneratedExclusions,
+    'Codacy exclusions must stay limited to generated/build/cache artifacts.'
+  );
+
+  for (const forbiddenPattern of forbiddenScopePatterns) {
+    assert.ok(
+      !exclusions.includes(forbiddenPattern),
+      `Codacy exclusions must keep ${forbiddenPattern} in analysis scope.`
+    );
+  }
 });
